Fix: if location editor is opened with selectTitle=true, the user can't change values in Blockly combo boxes
It appears that the select event is triggered even when the user just clicks around in the combo, if any text in the input is selected. The event triggers a state change and then re-rendering of the editor without first saving the new state of the combo.

This is just a hypothesis, but checking whether the input is actually focused fixed the bug.